# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales Manager API",
        "description": "This API manages sales operations, including consulting, creating, updating, and deleting clients and products, as well as querying sales data.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://0.0.0.0:8000"
        }
    ],
    "paths": {
...
        "/api/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get a list of all sales",
                "operationId": "419ebb078c83262b05f7c639ae8b2c21",
                "responses": {
                    "200": {
                        "description": "List of sales retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Sale"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Store a new sale",
                "operationId": "51f42c056032d143981f26f6c0888f24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale executed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sale executed successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Sale"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Product is not available or not enough stock"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Sale": {
                "title": "Sale",
                "description": "Sale model",
                "properties": {
                    "id": {
                        "description": "Sale ID",
                        "type": "integer"
                    },
                    "client_id": {
                        "description": "Client ID",
                        "type": "integer"
                    },
                    "product_id": {
                        "description": "Product ID",
                        "type": "integer"
                    },
                    "quantity": {
                        "description": "Quantity of products sold",
                        "type": "integer"
                    },
                    "unit_price": {
                        "description": "Unit price of the product",
                        "type": "number",
                        "format": "float"
                    },
                    "total_price": {
                        "description": "Total price of the sale",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}